outsourced open cart button as component
