+++ src/pageContent/store/cart/Cart.module.css
@@ -13,7 +13,7 @@
   text-align: center;
 }
 
-.cart button {
+.checkout {
   font-family: 'Orbitron', sans-serif;
   color: #fffdfd;
   border: 1px solid #fffdfd;
@@ -26,14 +26,14 @@
   transition: 0.5s ease;
 }
 
-.cart button:hover {
+.checkout:hover {
   color: #1110108f;
   border: 1px solid #1110108f;
   background: #fffdfd;
   transform: scale(1.1);
 }
 
-.cart button:disabled {
+.checkout:disabled {
   background: transparent;
   pointer-events: none;
 }
@@ -48,29 +48,6 @@
   height: 10rem;
   border-top: 1px solid hsla(0, 0%, 80%, 0.36);
   border-bottom: 1px solid hsla(0, 0%, 80%, 0.36);
-}
-
-/* styles for header button to open cart */
-.header {
-  display: flex;
-  margin: 0 1rem;
-}
-
-.header button {
-  color: #fffdfd;
-  background: none;
-  border: 1px solid transparent;
-  border-radius: 3px;
-  cursor: pointer;
-  position: relative;
-  z-index: 1;
-}
-
-.header p {
-  font-family: 'Orbitron', sans-serif;
-  font-size: larger;
-  margin: 0;
-  width: 2rem;
 }
 
 @media (max-width: 440px) {
